+++ jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    preset: 'jest-expo',
+    setupFilesAfterEnv: [
+        '@testing-library/jest-native/extend-expect',
+        './jest.setup.js'
+    ],
+    moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+};