+++ __tests__/App-test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import App from '../App';
 import { render, fireEvent, waitFor } from '@testing-library/react-native';
 
-import renderer from 'react-test-renderer';
+import renderer, {act} from 'react-test-renderer';
 import expect from "expect";
 
 test('renders correctly', () => {
@@ -20,14 +20,76 @@ beforeEach(() => {
     fetch.mockClear();
 });
 
-it('displays fetching message after button click', async () => {
+afterEach(() => {
+    jest.restoreAllMocks();
+});
+
+it('displays readiness message after button click', async () => {
     const { getByText } = render(<App />);
 
     // simulate button click
-    fireEvent.press(getByText('Readiness Check'));
+    await act(() => {
+        fireEvent.press(getByText('Readiness Check'));
+    });
 
     // check that the fetching message is displayed
     await waitFor(() => {
-        expect(getByText('Fetching data')).toBeDefined();
+        expect(getByText('Checking Readiness...')).toBeDefined();
+    });
+});
+
+it('displays liveness message after button click', async () => {
+    const { getByText } = render(<App />);
+
+    // simulate button click
+    await act(() => {
+        fireEvent.press(getByText('Liveness Check'));
+    });
+
+    // check that the fetching message is displayed
+    await waitFor(() => {
+        expect(getByText('Checking Liveness...')).toBeDefined();
+    });
+});
+
+it('displays version message after button click', async () => {
+    const { getByText } = render(<App />);
+
+    // simulate button click
+    await act(() => {
+        fireEvent.press(getByText('Version Check'));
+    });
+
+    // check that the fetching message is displayed
+    await waitFor(() => {
+        expect(getByText('Checking Version...')).toBeDefined();
+    });
+});
+
+it('displays verification message after button click', async () => {
+    const { getByText } = render(<App />);
+
+    // simulate button click
+    await act(() => {
+        fireEvent.press(getByText('Validate Blockchain'));
+    });
+
+    // check that the fetching message is displayed
+    await waitFor(() => {
+        expect(getByText('Checking Blockchain...')).toBeDefined();
+    });
+});
+
+it('displays block message after button click', async () => {
+    const { getByText } = render(<App />);
+
+    // simulate button click
+    await act(() => {
+        fireEvent.press(getByText('Add Block'));
+    });
+
+    // check that the fetching message is displayed
+    await waitFor(() => {
+        expect(getByText('Adding Block...')).toBeDefined();
     });
 });